+++ areas/VirtualDesktop/src/AzureMcp.VirtualDesktop/Commands/BaseVirtualDesktopCommand.cs
@@ -1,0 +1,34 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using AzureMcp.Areas.VirtualDesktop.Options;
+using AzureMcp.Core.Commands;
+using AzureMcp.Core.Commands.Subscription;
+using AzureMcp.Core.Options;
+
+namespace AzureMcp.Areas.VirtualDesktop.Commands;
+
+/// <summary>
+/// Base command for all Virtual Desktop commands
+/// </summary>
+public abstract class BaseVirtualDesktopCommand<
+    [DynamicallyAccessedMembers(TrimAnnotations.CommandAnnotations)] TOptions>
+    : SubscriptionCommand<TOptions>
+    where TOptions : SubscriptionOptions, new()
+{
+    protected new readonly Option<string> _resourceGroupOption = VirtualDesktopOptionDefinitions.ResourceGroup;
+
+    protected override void RegisterOptions(Command command)
+    {
+        base.RegisterOptions(command);
+        command.AddOption(_resourceGroupOption);
+    }
+
+    protected override TOptions BindOptions(ParseResult parseResult)
+    {
+        var options = base.BindOptions(parseResult);
+        options.ResourceGroup = parseResult.GetValueForOption(_resourceGroupOption);
+        return options;
+    }
+}